Allow restrictions based on Authorization Header bearer token hash value
In commit e1205b72b8 ("Allow restrictions based on Authorization header"),
we added restriction match type that tests the Auth Header of the
websocket upgrade request matches a regex.

There are cases where the wstunnel server cannot be provisioned with the
exact expected authorization value;
Let's provide a new MatchConfig type 'BearerHash', which expects a
Authorization header holding a bearer token, hashes it, and compares to
a given value.